crew-ai/requirements.txt
crewai
crewai_tools
load_dotenv
langchain_community
google-generativeai
ipython